1) Refactor CreateWallet.jsx: Move the Form component to CreateImportWalletForm.jsx and split the logic of importing and creating wallet into two separate containers (CreateWallet, ImportWallet) in the CreateImportWallet folder. 2) Extract loadWalletRecovery method from Setting class and make it exportable. 3) Add the option to recover a wallet in the ImportWallet container by reusing the loadWalletRecovery method.
