+++ app/containers/CreateImportWallet/CreateWallet/CreateWallet.scss
@@ -1,0 +1,8 @@
+.inputContainer {
+  height: 100%;
+}
+
+.iconDisplay {
+  display: flex;
+  margin-right: 10px;
+}
