+++ app/containers/CreateImportWallet/CreateWallet/index.js
@@ -1,0 +1,1 @@
+export { default } from './CreateWallet'
