+++ app/containers/CreateImportWallet/ImportWallet/ImportWallet.scss
@@ -1,0 +1,16 @@
+.iconDisplay {
+  display: flex;
+  margin-right: 10px;
+}
+
+.tabContainer {
+  display: flex;
+  flex-direction: column;
+  width: 500px;
+  align-items: center;
+}
+
+.buttonContainer {
+  height: 200px;
+  margin-bottom: 260px;
+}
