+++ app/containers/CreateImportWallet/ImportWallet/index.js
@@ -1,0 +1,40 @@
+// @flow
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
+import { compose } from 'recompose'
+import { withData, withActions, withCall } from 'spunky'
+import ImportWallet from './ImportWallet'
+import nodeStorageActions from '../../../actions/nodeStorageActions'
+import accountsActions, {
+  updateAccountsActions,
+} from '../../../actions/accountsActions'
+import {
+  showErrorNotification,
+  showSuccessNotification,
+} from '../../../modules/notifications'
+
+const actionCreators = {
+  showErrorNotification,
+  showSuccessNotification,
+}
+
+const mapDispatchToProps = dispatch =>
+  bindActionCreators(actionCreators, dispatch)
+
+const mapAccountsDataToProps = accounts => ({
+  accounts,
+})
+
+const mapAccountsActionsToProps = actions => ({
+  setAccounts: accounts => actions.call(accounts),
+})
+
+export default compose(
+  connect(
+    null,
+    mapDispatchToProps,
+  ),
+  withCall(nodeStorageActions),
+  withData(accountsActions, mapAccountsDataToProps),
+  withActions(updateAccountsActions, mapAccountsActionsToProps),
+)(ImportWallet)
